Need to update dependency-management-plugin plugin version
The versions of Gradle and Spring's dependency-management-plugin that are in use.

See the original report from the Spring guys in this Bug in Gradle 2.14-rc1 - No service of type StyledTextOutputFactory report(https://discuss.gradle.org/t/bug-in-gradle-2-14-rc1-no-service-of-type-styledtextoutputfactory/17638). Gradle moved the StyledTextOutputFactory to an internal package at some point (for the 3.0 release), which broke dependency-management-plugin 0.5.x.

This dependency-management-plugin issue(https://github.com/spring-gradle-plugins/dependency-management-plugin/issues/87) details their making changes to address this in their 0.6.0 release.

I see your build script references Gradle 2.3...but I'm wondering if that is accurate.

I think it boils down to either use Gradle 2.x with dependency-management-plugin 0.5.x or use Gradle 3.x with dependency-management-plugin 0.6.x.